requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.3
psycopg==3.1.17
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
